[vim] Add mappings to increase/decrease numbers
Increasing or decreasing the number currently below the cursor or on the
same line is by default done by pressing C-a or C-x. C-a interferes with
tmux, where it is the prefix key. These alternative mappings makes it
possible to increase and decrease numbers using <leader>-a and
<leader>-x.

More information:
- http://vim.wikia.com/wiki/Increasing_or_decreasing_numbers